+++ .gitignore
@@ -9,4 +9,3 @@ __pycache__
 coverage.xml
 .nox
 spec.json
-.idea
