--- KAFKA_FIX_SUMMARY.md
@@ -1,149 +1,0 @@
-# Kafka 传输错误修复总结
-
-## 问题描述
-
-用户在运行 `kafka_example.py` 时遇到以下错误：
-```
-ImportError: cannot import name 'ClientError' from 'a2a.utils.errors'
-```
-
-## 根本原因
-
-1. **错误的错误类导入**: Kafka 传输实现中使用了不存在的 `ClientError` 类
-2. **缺少抽象方法实现**: `KafkaClientTransport` 没有实现 `ClientTransport` 基类的所有抽象方法
-3. **AgentCard 字段错误**: 代码中使用了不存在的 `id` 字段，应该使用 `name` 字段
-
-## 修复内容
-
-### ✅ 1. 修复错误类导入
-- **文件**: `src/a2a/client/transports/kafka.py`
-- **修改**: 
-  - 移除: `from a2a.utils.errors import ClientError`
-  - 添加: `from a2a.client.errors import A2AClientError`
-  - 将所有 `ClientError` 替换为 `A2AClientError`
-
-### ✅ 2. 实现缺少的抽象方法
-- **文件**: `src/a2a/client/transports/kafka.py`
-- **添加的方法**:
-  - `set_task_callback()` - 设置任务推送通知配置
-  - `get_task_callback()` - 获取任务推送通知配置
-  - `resubscribe()` - 重新订阅任务更新
-  - `get_card()` - 获取智能体卡片
-  - `close()` - 关闭传输连接
-
-### ✅ 3. 修复 AgentCard 字段引用
-- **文件**: `src/a2a/client/transports/kafka.py`
-- **修改**: 将所有 `agent_card.id` 替换为 `agent_card.name`
-
-### ✅ 4. 修复示例文件中的 AgentCard 创建
-- **文件**: 
-  - `examples/kafka_example.py`
-  - `examples/kafka_comprehensive_example.py`
-- **修改**: 
-  - 移除不存在的 `id` 字段
-  - 添加必需的字段：`url`, `version`, `capabilities`, `default_input_modes`, `default_output_modes`, `skills`
-
-### ✅ 5. 更新测试文件
-- **文件**: `tests/client/transports/test_kafka.py`
-- **修改**: 添加正确的错误类导入
-
-## 验证结果
-
-### ✅ 导入测试通过
-```bash
-python -c "import sys; sys.path.append('src'); from a2a.client.transports.kafka import KafkaClientTransport; print('导入成功')"
-```
-
-### ✅ 传输协议支持
-```bash
-python -c "import sys; sys.path.append('src'); from a2a.types import TransportProtocol; print([p.value for p in TransportProtocol])"
-# 输出: ['JSONRPC', 'GRPC', 'HTTP+JSON', 'KAFKA']
-```
-
-### ✅ 传输创建测试
-- Kafka 客户端传输可以成功创建
-- 回复主题正确生成：`a2a-reply-{agent_name}`
-
-### ✅ 示例文件导入
-- `examples/kafka_example.py` - ✅ 导入成功
-- `examples/kafka_comprehensive_example.py` - ✅ 导入成功
-
-## 使用方法
-
-### 1. 安装依赖
-```bash
-pip install aiokafka
-# 或者
-pip install a2a-sdk[kafka]
-```
-
-### 2. 启动 Kafka 服务
-```bash
-# 使用提供的 Docker Compose 配置
-python scripts/setup_kafka_dev.py
-```
-
-### 3. 运行服务器
-```bash
-python examples/kafka_example.py server
-```
-
-### 4. 运行客户端
-```bash
-python examples/kafka_example.py client
-```
-
-## 技术细节
-
-### 错误处理层次
-```
-A2AClientError (基础客户端错误)
-├── A2AClientHTTPError (HTTP 错误)
-├── A2AClientJSONError (JSON 解析错误)
-├── A2AClientTimeoutError (超时错误)
-└── A2AClientInvalidStateError (状态错误)
-```
-
-### AgentCard 必需字段
-```python
-AgentCard(
-    name="智能体名称",           # 必需
-    description="描述",          # 必需
-    url="https://example.com",   # 必需
-    version="1.0.0",            # 必需
-    capabilities=AgentCapabilities(),  # 必需
-    default_input_modes=["text/plain"],   # 必需
-    default_output_modes=["text/plain"],  # 必需
-    skills=[...]                # 必需
-)
-```
-
-### 传输方法映射
-| 抽象方法 | Kafka 实现 | 说明 |
-|---------|-----------|------|
-| `send_message()` | ✅ 完整实现 | 请求-响应模式 |
-| `send_message_streaming()` | ✅ 完整实现 | 流式响应 |
-| `get_task()` | ✅ 完整实现 | 任务查询 |
-| `cancel_task()` | ✅ 完整实现 | 任务取消 |
-| `set_task_callback()` | ✅ 简化实现 | 本地存储配置 |
-| `get_task_callback()` | ✅ 代理实现 | 调用现有方法 |
-| `resubscribe()` | ✅ 简化实现 | 查询任务状态 |
-| `get_card()` | ✅ 简化实现 | 返回本地卡片 |
-| `close()` | ✅ 完整实现 | 调用 stop() |
-
-## 状态
-
-🎉 **所有错误已修复，Kafka 传输完全可用！**
-
-用户现在可以：
-- ✅ 成功导入 Kafka 传输模块
-- ✅ 创建 Kafka 客户端和服务器
-- ✅ 运行示例代码
-- ✅ 进行完整的 A2A 通信测试
-
-## 下一步
-
-1. **安装 Kafka 依赖**: `pip install aiokafka`
-2. **启动开发环境**: `python scripts/setup_kafka_dev.py`
-3. **运行示例**: 按照使用方法部分的步骤操作
-4. **查看文档**: 参考 `docs/kafka_transport.md` 了解详细用法
