+++ src/a2a/client/client_factory.py
@@ -110,8 +110,24 @@ class ClientFactory:
                 )
             self.register(
                 TransportProtocol.kafka,
-                KafkaClientTransport.create,
+                self._create_kafka_transport,
             )
+
+    def _create_kafka_transport(
+        self,
+        card: AgentCard,
+        url: str,
+        config: ClientConfig,
+        interceptors: list[ClientCallInterceptor],
+    ) -> ClientTransport:
+        """Create a Kafka transport that will auto-start when first used."""
+        # Create the transport using the existing create method
+        transport = KafkaClientTransport.create(card, url, config, interceptors)
+        
+        # Mark the transport for auto-start when first used
+        transport._auto_start = True
+        
+        return transport
 
     def register(self, label: str, generator: TransportProducer) -> None:
         """Register a new transport producer for a given transport label."""
