+++ src/a2a/server/apps/kafka/__init__.py
@@ -1,6 +1,6 @@
 """Kafka server application components for A2A."""
 
-from a2a.server.apps.kafka.app import KafkaServerApp
+from a2a.server.apps.kafka.kafka_app import KafkaServerApp
 
 __all__ = [
     'KafkaServerApp',
